+++ custom_components/titon_controller/const.py
@@ -1,0 +1,17 @@
+"""Constants for the Titon Controller integration."""
+
+DOMAIN = "titon_controller"
+
+CONF_SERIAL_PORT = "serial_port"
+CONF_WEB_HOST = "web_host"
+CONF_WEB_PORT = "web_port"
+CONF_PANEL_URL = "panel_url"
+CONF_HUMIDITY_SENSORS = "humidity_sensors"
+CONF_SETTINGS_PATH = "settings_path"
+CONF_LOG_PATH = "log_path"
+
+DEFAULT_SERIAL_PORT = "/dev/ttyUSB1"
+DEFAULT_WEB_HOST = "0.0.0.0"
+DEFAULT_WEB_PORT = 8050
+
+PANEL_URL_PATH = "titon-controller"